% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ask_toggl_api_token}
\alias{ask_toggl_api_token}
\title{ask_toggl_api_token}
\usage{
ask_toggl_api_token(msg = "toggl api token")
}
\arguments{
\item{msg}{the message}
}
\description{
ask for the toggle api token
}
