% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_id_to_name}
\alias{client_id_to_name}
\title{client_id_to_name}
\usage{
client_id_to_name(id, api_token = get_toggl_api_token())
}
\arguments{
\item{id}{client id}

\item{api_token}{the toggl api token}
}
\value{
the client name
}
\description{
return client name from client id
}
