% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_name_to_id}
\alias{client_name_to_id}
\title{client_name_to_id}
\usage{
client_name_to_id(name, api_token = get_toggl_api_token())
}
\arguments{
\item{name}{client name}

\item{api_token}{the toggl api token}
}
\value{
the client id
}
\description{
return client id from client name
}
