% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{correct_date}
\alias{correct_date}
\title{correct_date}
\usage{
correct_date(time)
}
\arguments{
\item{time}{as POSIXt}
}
\description{
bidouille pour avoir iso 8601, pas propre mais la tisuit, osef
}

