% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{create_client}
\alias{create_client}
\title{create_client}
\usage{
create_client(name = "client_demo", api_token = get_toggl_api_token(),
  wid = get_workspace_id())
}
\arguments{
\item{name}{client name}

\item{api_token}{the toggl api token}
}
\description{
create a client
}
\examples{
\dontrun{
get_toggl_api_token()# set your api token here
create_client("new client")
}
}
