% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clients.R
\name{createClient}
\alias{createClient}
\title{Create a client}
\usage{
createClient(workspace_id, name, notes = NULL)
}
\arguments{
\item{workspace_id}{workspace ID, where the client will be used (integer, required)}

\item{name}{The name of the client (string, required, unique in workspace)}

\item{notes}{Notes for the client (string, not required)}
}
\description{
Create a client
}
