% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{createProject}
\alias{createProject}
\title{Create project}
\usage{
createProject(workspace_id, name, client_id = NULL, is_private = TRUE)
}
\arguments{
\item{workspace_id}{workspace ID, where the project will be saved (integer, required)}

\item{name}{The name of the project (string, required, unique for client and workspace)}

\item{client_id}{client ID (integer, not required)}

\item{is_private}{whether project is accessible for only project users or for all workspace users (boolean, default true)}
}
\description{
Create project
}
