% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_training.R
\name{day_of_tranning}
\alias{day_of_tranning}
\title{Add day of training}
\usage{
day_of_tranning(name_tranning, day)
}
\arguments{
\item{name_tranning}{name  of compagny}

\item{day}{which day}
}
\description{
Add day of training
}
