% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_all_client_info}
\alias{get_all_client_info}
\title{get_all_client_info}
\usage{
get_all_client_info(api_token = get_toggl_api_token())
}
\arguments{
\item{api_token}{the toggl api token}
}
\value{
a data.frame
}
\description{
get_all_client_info
}
\examples{
\dontrun{
get_all_client_info()
}
}
