% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard.R
\name{get_all_client_names}
\alias{get_all_client_names}
\title{get all client's name}
\usage{
get_all_client_names(
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(),
  since = Sys.Date() - lubridate::years(1),
  until = Sys.Date()
)
}
\arguments{
\item{api_token}{the toggl api token}

\item{workspace_id}{the workspace id}

\item{since}{a date}

\item{until}{a date}
}
\description{
get all client's name
}
