% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_current}
\alias{get_current}
\title{get_current}
\usage{
get_current(api_token = get_toggl_api_token())
}
\arguments{
\item{api_token}{the toggl api token}
}
\description{
retrieve current projet id
}
