% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_duration.R
\name{get_current_duration}
\alias{get_current_duration}
\title{get current duration}
\usage{
get_current_duration(api_token = get_toggl_api_token())
}
\arguments{
\item{api_token}{the Token API}
}
\value{
a difftime
}
\description{
get current duration
}
\examples{
\dontrun{
get_current_duration()
}
}
