% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard.R
\name{get_dashboard}
\alias{get_dashboard}
\title{recupere les données depuis toggl.com}
\usage{
get_dashboard(api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(), since = Sys.Date() -
  lubridate::years(1), until = Sys.Date())
}
\arguments{
\item{api_token}{the toggl api token}

\item{workspace_id}{the workspace id}

\item{since}{begin date}

\item{until}{stop date}
}
\description{
recupere les données depuis toggl.com
}
