% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project_id}
\alias{get_project_id}
\title{get_project_id}
\usage{
get_project_id(project_name = get_context_project(),
  api_token = get_toggl_api_token(), create = FALSE)
}
\arguments{
\item{api_token}{the toggl api token}
}
\description{
retrieve projet id
}

