% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project_id_and_name}
\alias{get_project_id_and_name}
\title{get_project_id_and_name}
\usage{
get_project_id_and_name(
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token)
)
}
\arguments{
\item{api_token}{the toggl api token}

\item{workspace_id}{workspace id}
}
\description{
retrieve project id and name
}
