% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard.R
\name{get_project_task_detail}
\alias{get_project_task_detail}
\title{get all task from a project}
\usage{
get_project_task_detail(project_name = get_context_project(),
  api_token = get_toggl_api_token(), workspace_id = get_workspace_id(),
  since = Sys.Date() - lubridate::years(1), until = Sys.Date(),
  humain = TRUE)
}
\arguments{
\item{project_name}{project name}

\item{api_token}{the toggl api token}

\item{workspace_id}{the workspace id}

\item{since}{a date}

\item{until}{a date}

\item{humain}{boolean humain readable time}
}
\description{
get all task from a project
}
