% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard.R
\name{get_project_total}
\alias{get_project_total}
\title{get project total time}
\usage{
get_project_total(project_name = get_context_project(),
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token), since = Sys.Date() -
  lubridate::years(1), until = Sys.Date())
}
\arguments{
\item{project_name}{project name}

\item{api_token}{the toggl api token}

\item{workspace_id}{the workspace id}

\item{since}{a date}

\item{until}{a date}
}
\description{
get project total time
}
