% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_report.R
\encoding{UTF-8}
\name{get_summary_report}
\alias{get_summary_report}
\title{get summary report}
\usage{
get_summary_report(
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token),
  since = Sys.Date() - lubridate::years(1),
  until = Sys.Date(),
  user_agent = "togglr",
  users = get_workspace_users(api_token = api_token, workspace_id = workspace_id)
)
}
\arguments{
\item{api_token}{the toggl api token}

\item{workspace_id}{the workspace id}

\item{since}{begin date}

\item{until}{stop date}

\item{user_agent}{user_agent}

\item{users}{users}
}
\description{
by user then projet
}
