% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_time_entries.R
\encoding{UTF-8}
\name{get_time_entries}
\alias{get_time_entries}
\title{Get all time entries between 2 dates}
\usage{
get_time_entries(api_token = get_toggl_api_token(), since = Sys.Date() -
  lubridate::years(1), until = Sys.Date())
}
\arguments{
\item{since}{begin date}

\item{until}{stop date}
}
\value{
a data.frame containing all time entries
}
\description{
Get all time entries between 2 dates
}
\examples{
get_time_entries()
}
