% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_toggl_api_token}
\alias{get_toggl_api_token}
\title{get_toggl_api_token}
\usage{
get_toggl_api_token(ask = TRUE)
}
\arguments{
\item{ask}{booleen do we have to ask if missing}
}
\description{
return the toggle api token
}
