% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toggl_clients.R
\name{get_toggl_clients}
\alias{get_toggl_clients}
\title{Get Toggl Clients}
\usage{
get_toggl_clients(api_token, workspace_id, user_agent = "api_test")
}
\value{
a data frame of Toggl clients
}
\description{
Get Toggl Clients
}
