% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workspace.R
\name{get_workspace_id}
\alias{get_workspace_id}
\title{get_workspace_id}
\usage{
get_workspace_id(api_token = get_toggl_api_token())
}
\arguments{
\item{api_token}{the toggl api token}
}
\description{
Return the workspace id
}
