% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{getClients}
\alias{getClients}
\title{Get workspace clients}
\usage{
getClients(workspace_id)
}
\arguments{
\item{workspace_id}{}
}
\value{
data.table with columns: client_id, client_name
}
\description{
Get workspace clients
}
