% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{getProjects}
\alias{getProjects}
\title{Get workspace projects}
\usage{
getProjects(workspace_id)
}
\arguments{
\item{workspace_id}{}
}
\value{
data.table with columns: project_id, project_name, client_id
}
\description{
Get workspace projects
}
