% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification.R
\name{notification}
\alias{notification}
\title{show notification}
\usage{
notification(title, msg)
}
\arguments{
\item{title}{title of the notificatin}

\item{msg}{content of the notification}
}
\description{
show notification using message
}
