% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_toggl.R
\name{open_toggl_website_app}
\alias{open_toggl_website_app}
\title{open toggle.com website on timesheet page}
\usage{
open_toggl_website_app()
}
\description{
open toggle.com website on timesheet page
}
