% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_toggl.R
\name{open_toggl_website_profile}
\alias{open_toggl_website_profile}
\title{open toggle.com website on profile page}
\usage{
open_toggl_website_profile()
}
\description{
open toggle.com website on profile page
}
