% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_toggl_api_token}
\alias{set_toggl_api_token}
\title{set_toggl_api_token}
\usage{
set_toggl_api_token(token)
}
\arguments{
\item{token}{toggl api token}
}
\description{
set the toggle api token
}
