% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard.R
\name{to_humain}
\alias{to_humain}
\title{transforme time column into humain readable column}
\usage{
to_humain(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
transforme time column into humain readable column
}
