% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{toggl_create}
\alias{toggl_create}
\title{toggl_create}
\usage{
toggl_create(description = get_context(), start = now(), stop, duration,
  api_token = getOption("toggl_api_token"))
}
\arguments{
\item{description}{the task you did}

\item{start}{time in POSIXt}

\item{stop}{time in POSIXt}

\item{duration}{in seconds}

\item{api_token}{the toggl api token}
}
\description{
create a
}
\examples{
options(toggl_api_token = "XXXXXXXX")# set your toggl api token here
toggl_create(duration=1200)
}

