% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/put.R
\name{toggl_stop}
\alias{toggl_stop}
\title{toggl_stop}
\usage{
toggl_stop(api_token = getOption("toggl_api_token"))
}
\arguments{
\item{api_token}{the toggl api token}
}
\description{
stop the active task
}
\examples{
options(toggl_api_token = "XXXXXXXX")# set your api token here
toggl_start()
toggl_stop()
}

