% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{toggl_get_tags}
\alias{toggl_get_tags}
\alias{toggl_create_tag}
\alias{toggl_update_tag}
\alias{toggl_delete_tag}
\title{Toggle tags}
\usage{
toggl_get_tags(
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token)
)

toggl_create_tag(
  tag_name,
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token)
)

toggl_update_tag(tag_id, new_name, api_token = get_toggl_api_token())

toggl_delete_tag(tag_id, api_token = get_toggl_api_token())
}
\arguments{
\item{api_token}{the toggl api token}

\item{workspace_id}{workspace id}

\item{tag_name}{tag name}

\item{tag_id}{tag ID}

\item{new_name}{Updated tag name}
}
\description{
Tags (as ids and names) can be retrieved for a workspace (give its ID),
created, as well as updated and deleted (given its ID).
}
\examples{
\dontrun{
toggl_get_tags()
}

\dontrun{
toggl_create_tag("new tag")
}

\dontrun{
toggl_update_tag(1, "updated tag name")
}

\dontrun{
toggl_delete_tag(1)
}

}
