% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_update_time_entries.R
\name{toggl_update_entries}
\alias{toggl_update_entries}
\title{toggl_update_entries}
\usage{
toggl_update_entries(
  time_entry_ids,
  description = NULL,
  start = NULL,
  pid = NULL,
  stop = NULL,
  duration = NULL,
  tags = NULL,
  api_token = get_toggl_api_token()
)
}
\arguments{
\item{time_entry_ids}{id(s) of time entries to update}

\item{description}{the task you did}

\item{start}{time in POSIXt}

\item{pid}{pid}

\item{stop}{time in POSIXt}

\item{duration}{in seconds}

\item{tags}{tags}

\item{api_token}{the toggl api token}
}
\description{
update time entries
}
\examples{
\dontrun{
options(toggl_api_token = "XXXXXXXX")# set your toggl api token here
time_entry <- toggl_create(duration=1200)
time_entry_id <- content(time_entry)$data$id

toggl_update_entries( time_entry_id,
                    description = "new description",
                    duration = 100,
                    tags = c("tag1", "tag2"),
                    api_token=get_toggl_api_token())

}
}
