% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\name{togglr-package}
\alias{togglr-package}
\alias{togglr}
\title{This package allow user to track time  spend on R using the toggl.com timesheet}
\description{
toggl.com API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/togglr}
  \item Report bugs at \url{https://github.com/ThinkR-open/togglr/issues}
}

}
\author{
vincent <vincent@thinkr.fr>
}
