% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{create_client}
\alias{create_client}
\title{create_client}
\usage{
create_client(name = "wihtout client",
  api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token))
}
\arguments{
\item{name}{client name}

\item{api_token}{the toggl api token}

\item{workspace_id}{workspace id}
}
\description{
create a client
}
\examples{
\dontrun{
get_toggl_api_token()# set your api token here
create_client("new client")
}
}
