% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard.R
\encoding{UTF-8}
\name{get_dashboard}
\alias{get_dashboard}
\title{get all data from <http://toggl.com>}
\usage{
get_dashboard(api_token = get_toggl_api_token(),
  workspace_id = get_workspace_id(api_token), since = Sys.Date() -
  lubridate::years(1), until = Sys.Date())
}
\arguments{
\item{api_token}{the toggl api token}

\item{workspace_id}{the workspace id}

\item{since}{begin date}

\item{until}{stop date}
}
\description{
get all data from <http://toggl.com>
}
