% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project_id}
\alias{get_project_id}
\title{get_project_id}
\usage{
get_project_id(project_name = get_context_project(),
  api_token = get_toggl_api_token(), create = FALSE, client = NULL,
  workspace_id = get_workspace_id(api_token))
}
\arguments{
\item{project_name}{the project name}

\item{api_token}{the toggl api token}

\item{create}{boolean do you want to create the project if it doesnt exist ?}

\item{client}{client name}

\item{workspace_id}{workspace id}
}
\description{
retrieve project id
}
