% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_entries.R
\encoding{UTF-8}
\name{get_time_entries}
\alias{get_time_entries}
\title{Get all time entries between 2 dates}
\usage{
get_time_entries(api_token = get_toggl_api_token(), since = Sys.time()
  - lubridate::weeks(1), until = Sys.time())
}
\arguments{
\item{api_token}{the toggl api token}

\item{since}{begin date (One week ago by default)}

\item{until}{stop date (Now by defaut)}
}
\value{
a data.frame containing all time entries
}
\description{
Get all time entries between 2 dates
}
\examples{
\dontrun{
get_time_entries()
}
}
