% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_project.R
\name{toggl_create_project}
\alias{toggl_create_project}
\title{toggl_create_project}
\usage{
toggl_create_project(project_name = get_context_project(),
  api_token = get_toggl_api_token(), wid = get_workspace_id(),
  client = NULL)
}
\arguments{
\item{project_name}{project name}

\item{api_token}{the toggl api token#'}

\item{wid}{workspace_id}

\item{client}{client name}
}
\description{
create a project
}
\examples{
\dontrun{
toggl_create_project()
}
}
