% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{toggl_start}
\alias{toggl_start}
\title{toggl_start}
\usage{
toggl_start(description = get_context(), start = now(),
  api_token = getOption("toggl_api_token"))
}
\arguments{
\item{description}{the task you are doing}

\item{start}{start time in POSIXt}

\item{api_token}{the toggl api token}
}
\description{
start a task
}
\examples{
\dontrun{
options(toggl_api_token = "XXXXXXXX")# set your api token here
toggl_start()
}
}

