% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{toggl_start}
\alias{toggl_start}
\title{toggl_start}
\usage{
toggl_start(
  description = get_context(),
  client = "without client",
  project_name = get_context_project(),
  start = now(),
  api_token = get_toggl_api_token(),
  tags = NULL,
  workspace_id = get_workspace_id(api_token)
)
}
\arguments{
\item{description}{the task you are doing}

\item{client}{client name}

\item{project_name}{nom du projet}

\item{start}{start time in POSIXt}

\item{api_token}{the toggl api token}

\item{tags}{tags}

\item{workspace_id}{workspace id}
}
\description{
start a task
}
\examples{
\dontrun{
get_toggl_api_token()# set your api token here
toggl_start()
}
}
