% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{toggl_stop}
\alias{toggl_stop}
\title{toggl_stop}
\usage{
toggl_stop(current = get_current(), api_token = get_toggl_api_token())
}
\arguments{
\item{current}{list task id and start time}

\item{api_token}{the toggl api token}
}
\description{
stop the active task
}
\examples{
\dontrun{
options(toggl_api_token = "XXXXXXXX")# set your api token here
toggl_start()
toggl_stop()
}
}
