% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\name{togglr-package}
\alias{togglr-package}
\alias{togglr}
\title{This package allow user to track time  spend on R using the toggl.com timesheet}
\description{
This package allow user to track time  spend on R using the toggl.com timesheet
}
\author{
vincent <vincent@thinkr.fr>
}
